+++ CHANGELOG.md
@@ -37,6 +37,11 @@
 
 ### 0.23.1
 
+* Versions updates:
+    * Klock `1.8.6` -> `1.8.7`
+    * Ktor `1.3.0` -> `1.3.1`
+* Now it is possible to get updates by polling with custom executor engine
+
 ## 0.22.0
 
 * **`KtorCallFactory` must return `HttpStatement` instead of `HttpClientCall`**